Add SZ quantities to the lightcone code as part of the EAGLE extra I/O option

See merge request swift/swiftsim!1409

+++ src/cooling/COLIBRE/cooling.c
@@ -283,7 +283,7 @@ float cooling_get_temperature(const struct phys_const *phys_const,
  * @param p #part data.
  * @param xp Pointer to the #xpart data.
  */
-float cooling_get_electron_density(const struct phys_const *phys_const,
+double cooling_get_electron_density(const struct phys_const *phys_const,
                                    const struct hydro_props *hydro_props,
                                    const struct unit_system *us,
                                    const struct cosmology *cosmo,
